Update financial reporting timeline
We are on a 2 month delayed reporting system. CI provides us information for each month at the end of the following month, so the earliest we can report this information to the EC will be at their meeting 2 months later. I have included monthly and quarterly reporting dates, as well as our goal for the draft of the annual financial report. Please feel free to adjust the level of detail, remove the monthly reporting, etc.