IDC -> Trainers Leadership
Based on conversations with Trainers Leadership, I'm quite certain that they meant to refer to the Trainers Leadership Committee, not the Instructor Development Committee (which is what is linked here). 

I am submitting this as a PR to make it as easy as possible, but I also recognize that this may be text that can only be updated by the EC. Let me know if I should do anything in addition to submitting this PR here.